Revise Hierarchical blockchain figure
Co-authored-by: Kostis Karantias <kkarantias@gmail.com>

This new version of the figure clarifies three concerns which were clear
in the previous text, but not illustrated by the previous version of the
figure:

1. Every block points back to the genesis block. This has also been
mentioned in the figure label. This choice is important for security to
ensure superblocks have not been premined.

2. The direction of the blockchain is shown (arrows point backwards).

3. Blocks depicted at a higher level are equal to those below them. This
is made clear by using a symbol similar to an equals sign to connect
blocks.