Add disputeProof function
This removes the verification of the submited chain from the submission.
If an honest party wants to dispute a proof, that means that the proof
must be invalid at some point. The caller of the proof only needs to
indicate this point and doesn't need to search the whole proof for
inconsistencies. Of course, overall verification is supported using the
range arguments.